Some reorganization + review issues.
